# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.ConfigMapPatchRequestBody:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      namespace:
        type: string
    type: object
  controllers.DeploymentPatchRequestBody:
    properties:
      container:
        description: Deployment string `json:"deployment"`
        type: string
      image:
        type: string
      namespace:
        type: string
    type: object
  controllers.NSClonerRequestBody:
    properties:
      targetNamespace:
        description: SourceNamespace string `json:"sourceNamespace"`
        type: string
    type: object
  controllers.SecretPatchRequestBody:
    properties:
      data:
        additionalProperties: true
        type: object
      namespace:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Kubernetes Namespace Cloner API URI:<br>&nbsp;&nbsp;https://{nw-server-hostname}:8080/api/v1<br><br>
  title: Kubernetes Namespace Cloner API
  version: 3.0.0
paths:
  /configmaps/:configmap:
    post:
      consumes:
      - application/json
      description: Update a config map in a specific namespace
      parameters:
      - description: ConfigMap Update Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.ConfigMapPatchRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update a config map
  /deployments/:deployment:
    post:
      consumes:
      - application/json
      description: Update the image of a deployment in a specific namespace
      parameters:
      - description: Deployment Image Set Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.DeploymentPatchRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update deployment image
  /namespaces:
    get:
      description: Get all namespaces in the cluster
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: Get all namespaces
  /namespaces/:namespace/cloneNamespace:
    post:
      consumes:
      - application/json
      description: Clone a namespace and its objects to a new namespace
      parameters:
      - description: Namespace clone request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.NSClonerRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Clone a namespace
  /namespaces/:namespace/configmaps/display:
    get:
      description: Display all config maps in the specified namespace
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Display config maps for a specific namespace
  /namespaces/:namespace/deployments:
    get:
      description: Get all deployments in the specified namespace
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: Get deployments for a specific namespace
  /namespaces/:namespace/deployments/:deployment/scaledown:
    post:
      consumes:
      - application/json
      description: Decrease the replica count of a deployment in a specific namespace
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: Deployment name
        in: path
        name: deployment
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully scaled down the deployment
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Scale down a deployment
  /namespaces/:namespace/deployments/:deployment/scaleup:
    post:
      consumes:
      - application/json
      description: Increase the replica count of a deployment in a specific namespace
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: Deployment name
        in: path
        name: deployment
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully scaled up the deployment
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Scale up a deployment
  /namespaces/:namespace/deployments/display:
    get:
      description: Display all deployments in the specified namespace
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Display deployments for a specific namespace
  /namespaces/:namespace/secrets/display:
    get:
      description: Display all secrets in the specified namespace
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Display secrets for a specific namespace
  /namespaces/{namespace}/cronjobs/{cronJob}/scaledown:
    post:
      consumes:
      - application/json
      description: Decrease the number of instances for a specified CronJob in a given
        namespace
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob Name
        in: path
        name: cronJob
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully scaled down the CronJob
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Scale down a CronJob
      tags:
      - CronJob
  /namespaces/{namespace}/cronjobs/{cronJob}/scaleup:
    post:
      consumes:
      - application/json
      description: Increase the number of instances for a specified CronJob in a given
        namespace
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob Name
        in: path
        name: cronJob
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully scaled up the CronJob
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Scale up a CronJob
      tags:
      - CronJob
  /secrets/:secret:
    post:
      consumes:
      - application/json
      description: Update a secret in a specific namespace
      parameters:
      - description: Secret Update Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.SecretPatchRequestBody'
      - description: Secret patch request body
        in: body
        name: secretPatchRequestBody
        required: true
        schema:
          $ref: '#/definitions/controllers.SecretPatchRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update a secret
schemes:
- http
- https
swagger: "2.0"